docs: update feedback and pagination docs
