+++ docs/components/Feedback/Alert/Documentation.mdx
@@ -1,0 +1,22 @@
+import { Meta, Canvas } from '@storybook/blocks';
+
+import * as AlertStories from '../../../../src/components/feedback/Alert/Alert.stories';
+
+<Meta of={AlertStories} />
+
+{/* Documentation goes here */}
+
+# Alert
+
+This is the documentation for the Alert component
+
+{/* Documentation goes here */}
+
+<Canvas meta={AlertStories.default} />
+
+### Related Links
+
+| Type  | Resource                                                                                                                                               |
+| ----- | ------------------------------------------------------------------------------------------------------------------------------------------------------ |
+| Eames | [Alert Component](https://www.figma.com/design/veXnmignQnJz8StIq10VJ5/Eames-2.0---Foundations-%26-Components?node-id=413-15306&p=f&t=1Pi9thxrLJt7TsHN-0) |
+| AntD | [Alert Component](https://ant.design/components/alert) |
