+++ docs/components/Feedback/Notification/Documentation.mdx
@@ -1,0 +1,22 @@
+import { Meta, Canvas } from '@storybook/blocks';
+
+import * as NotificationStories from '../../../../src/components/feedback/Notification/Notification.stories';
+
+<Meta of={NotificationStories} />
+
+{/* Documentation goes here */}
+
+# Notification
+
+This is the documentation for the Notification component
+
+{/* Documentation goes here */}
+
+<Canvas meta={NotificationStories.default} />
+
+### Related Links
+
+| Type  | Resource                                                                                                                                               |
+| ----- | ------------------------------------------------------------------------------------------------------------------------------------------------------ |
+| Eames | [Notification Component](https://www.figma.com/design/veXnmignQnJz8StIq10VJ5/Eames-2.0---Foundations-%26-Components?node-id=422-13542&p=f&t=1Pi9thxrLJt7TsHN-0) |
+| AntD | [Notification Component](https://ant.design/components/notification) |
