+++ docs/components/Feedback/Popconfirm/Documentation.mdx
@@ -1,0 +1,22 @@
+import { Meta, Canvas } from '@storybook/blocks';
+
+import * as PopconfirmStories from '../../../../src/components/feedback/Popconfirm/Popconfirm.stories';
+
+<Meta of={PopconfirmStories} />
+
+{/* Documentation goes here */}
+
+# Popconfirm
+
+This is the documentation for the Popconfirm component
+
+{/* Documentation goes here */}
+
+<Canvas meta={PopconfirmStories.default} />
+
+### Related Links
+
+| Type  | Resource                                                                                                                                               |
+| ----- | ------------------------------------------------------------------------------------------------------------------------------------------------------ |
+| Eames | [Popconfirm Component](https://www.figma.com/design/veXnmignQnJz8StIq10VJ5/Eames-2.0---Foundations-%26-Components?node-id=422-13717&p=f&t=1Pi9thxrLJt7TsHN-0) |
+| AntD | [Popconfirm Component](https://ant.design/components/popconfirm) |
