+++ docs/components/Feedback/Progress/Documentation.mdx
@@ -1,0 +1,22 @@
+import { Meta, Canvas } from '@storybook/blocks';
+
+import * as ProgressStories from '../../../../src/components/feedback/Progress/Progress.stories';
+
+<Meta of={ProgressStories} />
+
+{/* Documentation goes here */}
+
+# Progress
+
+This is the documentation for the Progress component
+
+{/* Documentation goes here */}
+
+<Canvas meta={ProgressStories.default} />
+
+### Related Links
+
+| Type  | Resource                                                                                                                                               |
+| ----- | ------------------------------------------------------------------------------------------------------------------------------------------------------ |
+| Eames | [Progress Component](https://www.figma.com/design/veXnmignQnJz8StIq10VJ5/Eames-2.0---Foundations-%26-Components?node-id=422-13987&p=f&t=1Pi9thxrLJt7TsHN-0) |
+| AntD | [Progress Component](https://ant.design/components/progress) |
