+++ docs/components/Feedback/Result/Documentation.mdx
@@ -1,0 +1,22 @@
+import { Meta, Canvas } from '@storybook/blocks';
+
+import * as ResultStories from '../../../../src/components/feedback/Result/Result.stories';
+
+<Meta of={ResultStories} />
+
+{/* Documentation goes here */}
+
+# Result
+
+This is the documentation for the Result component
+
+{/* Documentation goes here */}
+
+<Canvas meta={ResultStories.default} />
+
+### Related Links
+
+| Type  | Resource                                                                                                                                               |
+| ----- | ------------------------------------------------------------------------------------------------------------------------------------------------------ |
+| Eames | [Result Component](https://www.figma.com/design/veXnmignQnJz8StIq10VJ5/Eames-2.0---Foundations-%26-Components?node-id=423-32&p=f&t=1Pi9thxrLJt7TsHN-0) |
+| AntD | [Result Component](https://ant.design/components/result) |
