+++ docs/components/Feedback/Skeleton/Documentation.mdx
@@ -1,0 +1,22 @@
+import { Meta, Canvas } from '@storybook/blocks';
+
+import * as SkeletonStories from '../../../../src/components/feedback/Skeleton/Skeleton.stories';
+
+<Meta of={SkeletonStories} />
+
+{/* Documentation goes here */}
+
+# Skeleton
+
+This is the documentation for the Skeleton component
+
+{/* Documentation goes here */}
+
+<Canvas meta={SkeletonStories.default} />
+
+### Related Links
+
+| Type  | Resource                                                                                                                                               |
+| ----- | ------------------------------------------------------------------------------------------------------------------------------------------------------ |
+| Eames | [Skeleton Component](https://www.figma.com/design/veXnmignQnJz8StIq10VJ5/Eames-2.0---Foundations-%26-Components?node-id=423-365&p=f&t=6nzqzvHIbyHgLbNh-0) |
+| AntD | [Skeleton Component](https://ant.design/components/skeleton) |
