+++ docs/components/Feedback/Spin/Documentation.mdx
@@ -1,0 +1,22 @@
+import { Meta, Canvas } from '@storybook/blocks';
+
+import * as SpinStories from '../../../../src/components/feedback/Spin/Spin.stories';
+
+<Meta of={SpinStories} />
+
+{/* Documentation goes here */}
+
+# Spin
+
+This is the documentation for the Spin component
+
+{/* Documentation goes here */}
+
+<Canvas meta={SpinStories.default} />
+
+### Related Links
+
+| Type  | Resource                                                                                                                                               |
+| ----- | ------------------------------------------------------------------------------------------------------------------------------------------------------ |
+| Eames | [Spin Component](https://www.figma.com/design/veXnmignQnJz8StIq10VJ5/Eames-2.0---Foundations-%26-Components?node-id=516-40776&p=f&t=CcQ8vx5kuy3kh5XE-0) |
+| AntD | [Spin Component](https://ant.design/components/spin) |
